sdk: extract owned store pipeline orchestration
